+++ src/scripts/main.js
@@ -34,6 +34,18 @@ const isLocalhost = function(hostname) {
 }
 
 /**
+ * Determines if user agent is a bot. Approach is to get most bots, assuming other bots don't run JS.
+ * Source: https://stackoverflow.com/questions/20084513/detect-search-crawlers-via-javascript/20084661
+ * @param {String} userAgent - User agent that should be tested.
+ * @returns {Boolean} isBot
+ */
+const isBot = function(userAgent) {
+
+	return (/bot|crawler|spider|crawling/i).test(userAgent)
+
+}
+
+/**
  * Gathers all platform-, screen- and user-related information.
  * @param {Boolean} detailed - Include personal data.
  * @returns {Object} attributes - User-related information.
@@ -142,8 +154,13 @@ const record = function(server, domainId, attrs, opts, active) {
 		return console.warn('Ackee ignores you because you are on localhost')
 	}
 
+<<<<<<< HEAD
 	if (document.cookie.indexOf('ackee_login=1') >= 0) {
 		return console.warn('Ackee ignores you because you are logged in')
+=======
+	if (isBot(navigator.userAgent) === true) {
+		return console.warn('Ackee ignores you because you are a bot')
+>>>>>>> develop
 	}
 
 	const url = endpoint(server)