+++ CHANGELOG.md
@@ -12,7 +12,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Changed
 
-- Dropped support for older Ackee versions
+- Dropped support for older Ackee versions (< 2.0)
+- `.record` now returns an object with a stop-function instead of returning the stop-function directly
 
 ## [3.3.0] - 2020-05-26