+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Changed
 
+- Recommend to use the script served by your Ackee installation (#23)
 - Use `window.outerWidth` and `window.outerHeight` to avoid that the browser triggers a repaint when the script loads (#12)
 
 ## [4.0.2] - 2020-09-20