+++ CHANGELOG.md
@@ -4,11 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [4.1.0] - 2020-11-03
+## [4.2.0] - 2020-11-15
 
 ### New
 
 - Ignore your own visits using the `ignoreOwnVisits` (thanks @yehudab)
+
+## [4.1.0] - 2020-11-03
+
 - Ignore bots (#25, #7, thanks @yehudab)
 
 ### Changed